move S_PlayerMovement to Scripts/Player and add namespace.
